Final website/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0
python-docx==0.8.11
fpdf2==2.7.6
pathlib==1.0.1
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pymupdf==1.23.14
docx2txt==0.8
